chore(endpoints): move each thread endpoint to a different file
