+++ src/openai_responses/__init__.py
@@ -2,12 +2,10 @@ from openai_responses.endpoints.assistants import AssistantsMock
 from openai_responses.endpoints.chat import ChatMock, ChatCompletionMock
 from openai_responses.endpoints.embeddings import EmbeddingsMock
 from openai_responses.endpoints.files import FilesMock
-from openai_responses.endpoints.threads import (
-    ThreadsMock,
-    MessagesMock,
-    RunsMock,
-    RunStepsMock,
-)
+from openai_responses.endpoints.threads import ThreadsMock
+from openai_responses.endpoints.messages import MessagesMock
+from openai_responses.endpoints.runs import RunsMock
+from openai_responses.endpoints.run_steps import RunStepsMock
 
 __all__ = [
     # main API
